Fixed bug causing 10-digit phone numbers to be incorrectly validated.

+++ app/src/main/java/com/example/eventlottery/Organizer/FacilityDetailsDialogueFragment.java
@@ -192,7 +192,7 @@ public class FacilityDetailsDialogueFragment extends DialogFragment {
             return Boolean.FALSE;
         }
         try {
-            Integer number = Integer.parseInt(str);
+            Long number = Long.parseLong(str);
             if (number < 0) {
                 return Boolean.FALSE;
             }
